refactor: move `ElementPerformFullscreenExit` struct to `fullscreen/element.rs`
