+++ components/script/dom/fullscreen/document.rs
@@ -14,7 +14,8 @@ use crate::dom::bindings::inheritance::Castable;
 use crate::dom::bindings::refcounted::{Trusted, TrustedPromise};
 use crate::dom::bindings::reflector::DomGlobal;
 use crate::dom::document::document::Document;
-use crate::dom::element::{Element, ElementPerformFullscreenEnter, ElementPerformFullscreenExit};
+use crate::dom::element::{Element, ElementPerformFullscreenEnter};
+use crate::dom::fullscreen::ElementPerformFullscreenExit;
 use crate::dom::node::NodeTraits;
 use crate::dom::promise::Promise;
 use crate::dom::types::HTMLDialogElement;
